e3_diffusion_for_molecules/requirements.txt
GitPython==3.1.43
Jinja2==3.1.4
KNN-CUDA==0.2
MarkupSafe==2.1.5
PyYAML==6.0.1
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
ase==3.23.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
compressed-tensors==0.6.0
contourpy==1.2.1
cycler==0.12.1
datasets==3.1.0
dgl==2.4.0+cu124
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
e2cnn==0.2.3
einops==0.8.0
equiadapt==0.1.1.post1.dev45+ge5e82b9
exceptiongroup==1.2.2
fastapi==0.115.4
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gguf==0.10.0
gitdb==4.0.11
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.7
igraph==0.11.8
imageio==2.34.2
importlib_metadata==8.5.0
interegular==0.3.3
jiter==0.7.0
joblib==1.4.2
jsonschema-specifications==2024.10.1
jsonschema==4.23.0
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
lark==1.2.2
lightning-utilities==0.11.6
llvmlite==0.43.0
lm-format-enforcer==0.10.6
matplotlib==3.7.0
mistral_common==1.4.4
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.54.3
opencv-python-headless==4.10.0.84
outlines==0.0.46
packaging==24.1
pandas==2.2.3
partial-json-parser==0.2.1.1.post4
pillow==10.4.0
pip==24.2
platformdirs==4.2.2
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
protobuf==5.27.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==18.0.0
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pynauty==2.8.8.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytz==2024.2
pyzmq==26.2.0
ray==2.38.0
rdkit==2024.3.5
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.21.0
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.12.0
sentencepiece==0.2.0
sentry-sdk==2.12.0
setproctitle==1.3.3
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.1
texttable==1.7.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.12
tokenizers==0.20.3
torch-fidelity==0.3.0
torch==2.4.0
torch_geometric==2.5.3
torcheval==0.0.7
torchmetrics==1.4.1
torchsummary==1.5.1
torchvision==0.19.0
tqdm==4.66.4
transformers==4.46.2
trimesh==4.5.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
uvicorn==0.32.0
uvloop==0.21.0
vllm==0.6.3.post1
wandb==0.17.5
watchfiles==0.24.0
websockets==13.1
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.9.4
zipp==3.20.2